scripts/gh_digest/requirements.txt
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
openai>=1.40.0
